Rewrite upper: 46.200235 s val tupper : Bound.Sigs.Q.coef Bound.Sigs.Expr.expr =   2451785645616669/9007199254740992 Rewrite lower: 59.089219 s val tlower : Bound.Sigs.Q.coef Bound.Sigs.Expr.expr =   4964275859527245/9007199254740992
Interestingly, it is possible that non-integer values of x could require more iterations than are required to converge for the integer values. The following is an indication (but works with the first commented out version of a Newton iteration, that uses two floors, not just one --- see the previous commit). However, this is not proof that the upper bound we're computing here is not tight, because the error we're checking is (res-1)^2-x, and here it's res^2-x.
>>> def isqrt(x, r0, n):
...     for _ in xrange(0,n):
...         r0 = (r0 + x / r0) / 2
...         return r0
...
>>> def err(x, r0, n):
...     return (isqrt(x, r0, n))**2 - x
...
>>> err(1,2,2)
0
>>> err(2,2,2)
-1
>>> err(3,2,2)
-2
>>> err(4,2,2)
0
>>> err(1.5,2,2)
0.390625